ks_feed_proj/ks_feedlib/pages/com_kuaishou_live_core_basic_activity_LiveSlideActivity.py
# -*- coding: utf-8 -*-
import time

from uibase.upath import *
from uibase.controls import Window


class HomeActivity(Window):
    window_spec = {"path": UPath(activity_ == 'com.yxcorp.gifshow.HomeActivity')}

    def get_locators(self):
        return {
        }


