ks_feed_proj/requirements.txt
shoots>=0.7.3
shoots-uibase
shoots-android
shoots-android-byted
ui-ext-byted